src/requirements.txt
accelerate==0.33.0
beautifulsoup4==4.12.3
cpu_heater==0.1.0
editdistance==0.8.1
fire==0.6.0
fuzzywuzzy==0.18.0
GitPython==3.1.43
httpx==0.27.0
icecream==2.1.3
networkx==3.3
notion2pandas==1.0.0
notion_client==2.2.1
numpy==2.0.1
nvdutils>=0.2.1
openai==1.37.1
pandas==2.2.2
peft==0.12.0
PyDriller==1.15.5
PyGithub==2.3.0
Levenshtein==0.25.1
redis==3.5.3
regex==2024.5.15
requests==2.32.3
sympy==1.12
tiktoken==0.7.0
torch==2.3.1
tqdm>=4.66.4
transformers==4.42.4
tree_sitter>=0.23.0
tree_sitter_c>=0.23.0
tree_sitter_cpp>=0.23.0
tree_sitter_java>=0.23.2
unidiff==0.6.0
codebleu==0.7.0
pygraphviz>=1.13
thefuzz==0.22.1
